Consolidate transport types into single file
