--- packages/core/src/interfaces/common/net/index.ts
@@ -1,3 +1,0 @@
-export * from './rpc'
-export * from './transport'
-export * from './adapter.interface'
