--- packages/core/src/interfaces/common/net/rpc/index.ts
@@ -1,2 +1,0 @@
-export * from './rpc.interface'
-export * from './json-rpc'
