--- packages/core/src/interfaces/common/net/rpc/json-rpc/index.ts
@@ -1,2 +1,0 @@
-export * from './json-rpc-message.interface'
-export * from './json-rpc-adapter.interface'
