--- packages/core/src/interfaces/common/net/rpc/json-rpc/json-rpc-adapter.interface.ts
@@ -1,12 +1,0 @@
-import { RpcTransportAdapter } from '../../adapter.interface'
-import { JsonRpcRequest, JsonRpcResponse } from './json-rpc-message.interface'
-
-export type JsonRpcAdapter<
-  TransportRequest,
-  TransportResponse
-> = RpcTransportAdapter<
-  JsonRpcRequest,
-  JsonRpcResponse,
-  TransportRequest,
-  TransportResponse
->
