--- packages/core/src/interfaces/common/net/transport/http/http-transport.interface.ts
@@ -1,6 +1,0 @@
-/* Internal Imports */
-import { Client, Server } from '../transport.interface'
-import { HttpRequest, HttpResponse } from './http-message.interface'
-
-export type HttpClient = Client<HttpRequest, HttpResponse>
-export type HttpServer = Server
