--- packages/core/src/interfaces/common/net/transport/index.ts
@@ -1,3 +1,0 @@
-export * from './http/http-message.interface'
-export * from './http/http-transport.interface'
-export * from './transport.interface'
