--- packages/core/src/interfaces/transport/index.ts
@@ -1,4 +1,0 @@
-export interface ServerBackend {
-  app: any
-  listen(): Promise<void>
-}
